fix(kas-fleet-manager): ensure that the cluster status is cluster_provisioned
kas-installer relied on a soon to be removed KFM feature that resetted the standalone cluster status
to "cluster_provisioning" to allow for terraforming phase to take place. That's why it was okay to
set the status of standalone cluster to "ready" as it was reset to "provisioning" no matter what.
But this KFM capability didn't take onto account that someone can have an already terraformed
cluster and set their status to "ready" meaning that KFM doesn't have to auto terraform the cluster.
With the change in https://github.com/bf2fc6cc711aee1a0c2a/kas-fleet-manager/pull/1355 and notably
the
https://github.com/bf2fc6cc711aee1a0c2a/kas-fleet-manager/pull/1355/commits/acdaa4db9f39369fbe799b75ee05311a8c72d27d
commit, KFM wont change the status if they are different from "accepted" so kas-installer has to be
updated to reflect for this.